Renamed //hadoop to //mapreduce
