+++ server/resources/default-OLAP-configs.properties
@@ -39,7 +39,7 @@ cassandra.input.partitioner.class=org.apache.cassandra.dht.Murmur3Partitioner
 # Class that gremlin injects as a graph on which to run analytics
 gremlin.graph=org.apache.tinkerpop.gremlin.hadoop.structure.HadoopGraph
 # Input API implementation on the graph. If the janus or graph backend is changed, this should change accordingly.
-gremlin.hadoop.graphReader=grakn.core.hadoop.InputFormatHadoop
+gremlin.hadoop.graphReader=grakn.core.mapreduce.InputFormatHadoop
 # Output API implementation on the graph
 gremlin.hadoop.graphWriter=org.apache.tinkerpop.gremlin.hadoop.structure.io.gryo.GryoOutputFormat
 # If the classpath has been provided on the remote machine
